module-1-stateful-agent/requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
tavily-python==0.5.0
python-dotenv==1.0.0
pydantic==2.10.3
